Moved A_Markdown apart.
