+++ src/Ampersand/Basics.hs
@@ -3,6 +3,7 @@ module Ampersand.Basics
   , module Ampersand.Basics.Collection
   , module Ampersand.Basics.Exit
   , module Ampersand.Basics.Languages
+  , module Ampersand.Basics.PandocExtended
   , module Ampersand.Basics.String
   , module Ampersand.Basics.Unique
   , module Ampersand.Basics.UTF8
@@ -12,6 +13,7 @@ import Ampersand.Basics.Auxiliaries
 import Ampersand.Basics.Collection
 import Ampersand.Basics.Exit
 import Ampersand.Basics.Languages
+import Ampersand.Basics.PandocExtended
 import Ampersand.Basics.String
 import Ampersand.Basics.Unique
 import Ampersand.Basics.UTF8
