+++ src/Ampersand/Basics/PandocExtended.hs
@@ -1,13 +1,27 @@
-module Ampersand.Misc.Explain
-    ( string2Blocks
-    , blocks2String
-    , PandocFormat(..)
-    )
-where
+{-# LANGUAGE DeriveDataTypeable #-}
+module Ampersand.Basics.PandocExtended
+   ( PandocFormat(..)
+   , A_Markup(..)
+   , aMarkup2String
+   , string2Blocks
+   )
+where 
 
-import Text.Pandoc
-import Ampersand.Core.ParseTree      (PandocFormat(..))
-import Ampersand.Basics
+import Data.Typeable
+import Data.Data
+import Text.Pandoc hiding (Meta)
+import Ampersand.Basics.Languages
+import Ampersand.Basics.Version
+
+data PandocFormat = HTML | ReST | LaTeX | Markdown deriving (Eq, Show, Ord)
+
+data A_Markup =
+    A_Markup { amLang :: Lang -- No Maybe here!  In the A-structure, it will be defined by the default if the P-structure does not define it. In the P-structure, the language is optional.
+             , amPandoc :: [Block]
+             } deriving (Show, Eq, Prelude.Ord, Typeable, Data)
+
+aMarkup2String :: PandocFormat -> A_Markup -> String
+aMarkup2String fmt a = blocks2String fmt False (amPandoc a)
 
 -- | use a suitable format to read generated strings. if you have just normal text, ReST is fine.
 -- | defaultPandocReader getOpts should be used on user-defined strings.
