+++ src/Ampersand/Core/AbstractSyntaxTree.hs
@@ -30,7 +30,6 @@ module Ampersand.Core.AbstractSyntaxTree (
  , Expression(..)
  , getExpressionRelation
  , A_Concept(..)
- , A_Markup(..)
  , AMeaning(..)
  , A_RoleRule(..)
  , A_RoleRelation(..)
@@ -44,7 +43,6 @@ module Ampersand.Core.AbstractSyntaxTree (
  , ContextInfo(..)
  , showValADL,showValPHP,showValSQL
  , showSign
- , aMarkup2String
 -- , module Ampersand.Core.ParseTree  -- export all used constructors of the parsetree, because they have actually become part of the Abstract Syntax Tree.
  , (.==.), (.|-.), (./\.), (.\/.), (.-.), (./.), (.\.), (.<>.), (.:.), (.!.), (.*.)
  , makeConcept
@@ -56,7 +54,6 @@ import Ampersand.Core.ParseTree
     , PairView(..),PairViewSegment(..),Prop(..),Lang, PandocFormat
     , Representation(..), TType(..), PAtomValue(..), PSingleton
     )
-import Ampersand.Misc
 import Text.Pandoc hiding (Meta)
 import Data.Function
 import Data.Typeable
@@ -141,11 +138,6 @@ data A_RoleRule = A_RoleRule { arRoles :: [Role]
                              , arRules ::  [String] -- the names of the rules
                              , arPos ::   Origin
                              } deriving (Show)
-data A_Markup =
-    A_Markup { amLang :: Lang -- No Maybe here!  In the A-structure, it will be defined by the default if the P-structure does not define it. In the P-structure, the language is optional.
-             , amPandoc :: [Block]
-             } deriving (Show, Eq, Prelude.Ord, Typeable, Data)
-
 data RuleOrigin = UserDefined     -- This rule was specified explicitly as a rule in the Ampersand script
                 | Multiplicity    -- This rule follows implicitly from the Ampersand script (Because of a property) and generated by a computer
                 | Identity        -- This rule follows implicitly from the Ampersand script (Because of a identity) and generated by a computer
@@ -271,9 +263,6 @@ showDcl forceBoth dcl = name dcl++"["++cpts++"]"
     cpts
      | forceBoth || source dcl /= target dcl = show (source dcl) ++ "*"++ show (target dcl)
      | otherwise                             = show (source dcl)
-
-aMarkup2String :: PandocFormat -> A_Markup -> String
-aMarkup2String fmt a = blocks2String fmt False (amPandoc a)
 
 data AMeaning = AMeaning { ameaMrk ::[A_Markup]} deriving (Show, Eq, Prelude.Ord, Typeable, Data)
