+++ src/Ampersand/FSpec/GenerateUML.hs
@@ -1,11 +1,10 @@
 module Ampersand.FSpec.GenerateUML (generateUML) where
 
 import Ampersand.Basics
-import Ampersand.Core.AbstractSyntaxTree (explMarkup,aMarkup2String,Rule,Purpose(..))
+import Ampersand.Core.AbstractSyntaxTree (explMarkup,Rule,Purpose(..))
 import Ampersand.Graphic.ClassDiagram
 import Ampersand.Graphic.Fspec2ClassDiagrams 
 import Ampersand.FSpec
-import Ampersand.Core.ParseTree(PandocFormat(ReST))
 import Ampersand.Core.AbstractSyntaxTree
      ( Declaration )
 import Data.List
@@ -172,7 +171,9 @@ genCustomProfileElements =
     reqUML :: ReqValue2 -> String
     reqUML (xmiId, req) = intercalate "\n"
      ( ["   <thecustomprofile:Functional base_Requirement="++show xmiId++"/>"]++
-       [tagUML xmiId count puprtxt reftxt | (count, (puprtxt, reftxt)) <- zip [0::Int ..] [(aMarkup2String ReST (explMarkup p), intercalate ";" (explRefIds p)) | p <- reqPurposes req]]
+       [tagUML xmiId count puprtxt reftxt 
+       | (count, (puprtxt, reftxt)) <- zip [0::Int ..] [(aMarkup2String ReST (explMarkup p), intercalate ";" (explRefIds p)) | p <- reqPurposes req]
+       ]
      )
     tagUML xmiId nr value reftxt = intercalate "\n"
       [ "     <thecustomprofile:"++keyMeaning++" base_Requirement="++show xmiId++" "++keyMeaning++"="++show value++"/>"
