+++ src/Ampersand/Misc.hs
@@ -1,6 +1,5 @@
 module Ampersand.Misc
    ( module Ampersand.Misc.Options
-   , module Ampersand.Misc.Explain
    ) where
 import Ampersand.Misc.Options
         ( Options(..)
@@ -12,5 +11,3 @@ import Ampersand.Misc.Options
         , helpNVersionTexts
         , writeConfigFile
         )
-import Ampersand.Misc.Explain
-       (string2Blocks, blocks2String, PandocFormat(..))
